colores de error por fuera de los valores

+++ cd2024bfi1g1-model/src/main/resources/db/migration/V11__create_min_max_temp.sql
@@ -1,2 +1,2 @@
-Alter TABLE public.lots ADD min_temp int4;
-Alter TABLE public.lots ADD max_temp int4;
+Alter TABLE public.lots ADD min_temp float8;
+Alter TABLE public.lots ADD max_temp float8;
